Split ICDex client code into its own library (#4318)
