+++ Cargo.toml
@@ -98,6 +98,7 @@ members = [
     "backend/libraries/http_request",
     "backend/libraries/human_readable",
     "backend/libraries/human_readable_derive",
+    "backend/libraries/icdex_client",
     "backend/libraries/index_store",
     "backend/libraries/ledger_utils",
     "backend/libraries/msgpack",
