+++ backend/libraries/icdex_client/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "icdex_client"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+candid = { workspace = true }
+canister_client = { path = "../canister_client" }
+hex = { workspace = true }
+ic-cdk = { workspace = true }
+icrc1_ledger_canister_c2c_client = { path = "../../external_canisters/icrc1_ledger/c2c_client" }
+serde = { workspace = true }
+types = { path = "../types" }
