+++ backend/libraries/types/src/lib.rs
@@ -25,6 +25,7 @@ mod event_index;
 mod event_result;
 mod event_wrapper;
 mod events;
+mod exchanges;
 mod field_invalid;
 mod file;
 mod file_status;
@@ -86,6 +87,7 @@ pub use event_index::*;
 pub use event_result::*;
 pub use event_wrapper::*;
 pub use events::*;
+pub use exchanges::*;
 pub use field_invalid::*;
 pub use file::*;
 pub use file_status::*;
